[FLINK-29846] Add ITCase suffix for hadoop tests extending AbstractTestBase, allow empty for Minicluster extension rule
