--- flink-formats/flink-hadoop-bulk/archunit-violations/83371291-f688-4eaf-a207-24981f1067f3
@@ -1,3 +1,0 @@
-simple name of org.apache.flink.formats.hadoop.bulk.HadoopPathBasedPartFileWriterTest does not end with 'ITCase' in (HadoopPathBasedPartFileWriterTest.java:0)
-simple name of org.apache.flink.formats.hadoop.bulk.committer.HadoopRenameCommitterHDFSTest does not end with 'ITCase' in (HadoopRenameCommitterHDFSTest.java:0)
-simple name of org.apache.flink.formats.hadoop.bulk.committer.HadoopRenameCommitterLocalFSTest does not end with 'ITCase' in (HadoopRenameCommitterLocalFSTest.java:0)
